+++ packages/deploy-config-generator-shared/src/utils/constants.ts
@@ -3,3 +3,8 @@ export const cdsPkg = '@sap/cds-dk';
 export const mtaExecutable = 'mta';
 export const mtaPkg = 'mta';
 export const mtaYaml = 'mta.yaml';
+
+export enum TargetName {
+    ABAP = 'abap',
+    CF = 'cf'
+}