+++ packages/preview-middleware/test/unit/base/flp.test.ts
@@ -240,6 +240,15 @@ describe('FlpSandbox', () => {
             await flp.init(manifest);
             expect(flp.templateConfig.ui5.libs).toMatchSnapshot();
         });
+
+        test('do not add component to applications', async () => {
+            const manifest = {
+                'sap.app': { id: 'my.id', type: 'component' }
+            } as Manifest;
+            const flp = new FlpSandbox({}, mockProject, mockUtils, logger);
+            await flp.init(manifest);
+            expect(flp.templateConfig.apps).toMatchSnapshot();
+        });
     });
 
     describe('router', () => {