+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/odata-cli
 
+## 0.16.5
+
+### Patch Changes
+
+-   Updated dependencies [65f15d9]
+    -   @sap-ux/btp-utils@1.0.1
+    -   @sap-ux/axios-extension@1.18.5
+
 ## 0.16.4
 
 ### Patch Changes