+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.8.33
+
+### Patch Changes
+
+-   Updated dependencies [65f15d9]
+    -   @sap-ux/btp-utils@1.0.1
+    -   @sap-ux/axios-extension@1.18.5
+
 ## 0.8.32
 
 ### Patch Changes