+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/inquirer-common
 
+## 0.6.14
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.27
+-   @sap-ux/telemetry@0.5.59
+
+## 0.6.13
+
+### Patch Changes
+
+-   Updated dependencies [65f15d9]
+    -   @sap-ux/btp-utils@1.0.1
+    -   @sap-ux/fiori-generator-shared@0.7.26
+    -   @sap-ux/telemetry@0.5.58
+
 ## 0.6.12
 
 ### Patch Changes