+++ packages/nodejs-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/nodejs-utils
 
+## 0.1.7
+
+### Patch Changes
+
+-   Updated dependencies [65f15d9]
+    -   @sap-ux/btp-utils@1.0.1
+
 ## 0.1.6
 
 ### Patch Changes