+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 0.0.57
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.27
+-   @sap-ux/ui5-library-writer@0.5.48
+-   @sap-ux/ui5-library-inquirer@0.3.43
+
+## 0.0.56
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.26
+-   @sap-ux/nodejs-utils@0.1.7
+-   @sap-ux/ui5-library-inquirer@0.3.42
+
 ## 0.0.55
 
 ### Patch Changes